Moving the following definitions following review:
* AxisType
* Handedness
* RotationConstraintType

Also removing the Controller definition as it was replaced with an interface.
