--- Assets/MixedRealityToolkit/_Core/Definitions/Devices/Controller.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 4e10c37cac392d442ae96b3da1ab6b07
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {fileID: 2800000, guid: 8a1b00ef1a7a8b94891466c1f0911027, type: 3}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
