+++ Assets/MixedRealityToolkit/_Core/Definitions/Utilities/Handedness.cs
@@ -3,7 +3,7 @@
 
 using System;
 
-namespace Microsoft.MixedReality.Toolkit.Internal.Definitions
+namespace Microsoft.MixedReality.Toolkit.Internal.Definitions.Utilities
 {
     /// <summary>
     /// The Handedness defines which hand a controller is currently operating in.
