+++ Assets/MixedRealityToolkit/_Core/Utilities/Physics/TwoHandRotateLogic.cs
@@ -1,7 +1,7 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See LICENSE in the project root for license information.
 
-using Microsoft.MixedReality.Toolkit.Internal.Definitions;
+using Microsoft.MixedReality.Toolkit.Internal.Definitions.Utilities;
 using System;
 using System.Collections.Generic;
 using UnityEngine;
